requirements.txt
# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# OCR engines
pytesseract>=0.3.10
paddleocr>=2.6.0
paddlepaddle>=2.4.0

# PDF processing
pdf2image>=1.16.0
PyPDF2>=3.0.0

# Vision-language models
openai>=1.0.0
anthropic>=0.7.0

# Layout detection
layoutparser>=0.3.4
# detectron2 - installation may require specific instructions for your system
# Install with: python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'

# Image processing
scikit-image>=0.19.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0

# Data processing
pandas>=1.3.0

# Optional: Advanced ML libraries
torch>=1.13.0
torchvision>=0.14.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
